refactor: Migrate font-weight to treat
