+++ lib/components/Strong/Strong.tsx
@@ -1,12 +1,10 @@
 import React, { ReactNode } from 'react';
-import { useTheme } from '../private/ThemeContext';
+import { useWeight } from '../../hooks/useTypography';
 
 export interface StrongProps {
   children: ReactNode;
 }
 
-export const Strong = ({ children }: StrongProps) => {
-  const theme = useTheme();
-
-  return <strong className={theme.atoms.fontWeight.strong}>{children}</strong>;
-};
+export const Strong = ({ children }: StrongProps) => (
+  <strong className={useWeight('strong')}>{children}</strong>
+);
