+++ lib/hooks/useTypography/index.treat.ts
@@ -1,5 +1,17 @@
-import { style } from 'sku/treat';
+import { style, css } from 'sku/treat';
 
 export const fontFamily = style(({ typography }) => ({
   fontFamily: typography.fontFamily,
+}));
+
+export const fontWeight = css(({ typography }) => ({
+  regular: {
+    fontWeight: typography.fontWeight.regular,
+  },
+  medium: {
+    fontWeight: typography.fontWeight.medium,
+  },
+  strong: {
+    fontWeight: typography.fontWeight.strong,
+  },
 }));
