+++ lib/hooks/useTypography/index.ts
@@ -1,4 +1,14 @@
 import { useClassNames } from 'sku/treat';
-import { fontFamily } from './index.treat';
+import { fontFamily, fontWeight } from './index.treat';
 
-export default () => useClassNames(fontFamily);
+export type FontWeight = 'regular' | 'medium' | 'strong';
+
+interface Type {
+  weight?: FontWeight;
+}
+
+export default ({ weight }: Type) =>
+  useClassNames(fontFamily, useWeight(weight));
+
+export const useWeight = (weight: Type['weight'] = 'regular') =>
+  useClassNames(fontWeight[weight]);
