+++ lib/themes/jobStreet/atoms.css.js
@@ -47,12 +47,6 @@ const backgroundColor = {
   positive,
 };
 
-const fontWeights = {
-  regular: 400,
-  medium: 500,
-  strong: 600,
-};
-
 const transforms = {
   touchable: 'scale(0.98)',
 };
@@ -62,7 +56,6 @@ export default makeAtoms(
   colors,
   borderRadius,
   boxShadows,
-  fontWeights,
   backgroundColor,
   transforms,
 );
