+++ lib/themes/seekAnz/atoms.css.js
@@ -46,12 +46,6 @@ const backgroundColor = {
   positive,
 };
 
-const fontWeights = {
-  regular: 400,
-  medium: 500,
-  strong: 600,
-};
-
 const transforms = {
   touchable: 'scale(0.95)',
 };
@@ -61,7 +55,6 @@ export default makeAtoms(
   colors,
   borderRadius,
   boxShadows,
-  fontWeights,
   backgroundColor,
   transforms,
 );
