+++ lib/themes/seekAsia/atoms.css.js
@@ -56,12 +56,6 @@ const backgroundColor = {
   positive,
 };
 
-const fontWeights = {
-  regular: 400,
-  medium: 500,
-  strong: 700,
-};
-
 const transforms = {
   touchable: 'scale(0.98)',
 };
@@ -71,7 +65,6 @@ export default makeAtoms(
   colors,
   borderRadius,
   boxShadows,
-  fontWeights,
   backgroundColor,
   transforms,
 );
