+++ lib/themes/seekAsia/atoms.css.js.d.ts
@@ -55,9 +55,6 @@ export const fontSize_level1: string;
 export const fontSize_level2: string;
 export const fontSize_level3: string;
 export const fontSize_standard: string;
-export const fontWeight_medium: string;
-export const fontWeight_regular: string;
-export const fontWeight_strong: string;
 export const marginBottomDesktop_large: string;
 export const marginBottomDesktop_medium: string;
 export const marginBottomDesktop_none: string;
