+++ lib/themes/wireframe/atoms.css.js
@@ -45,12 +45,6 @@ const backgroundColor = {
   positive,
 };
 
-const fontWeights = {
-  regular: 400,
-  medium: 500,
-  strong: 600,
-};
-
 const transforms = {
   touchable: 'scale(0.97)',
 };
@@ -60,7 +54,6 @@ export default makeAtoms(
   colors,
   borderRadius,
   boxShadows,
-  fontWeights,
   backgroundColor,
   transforms,
 );
