+++ lib/themes/wireframe/theme.treat.ts
@@ -11,6 +11,11 @@ export default makeTreatTheme({
   name: 'wireframe',
   typography: {
     fontFamily: 'Courier, monospace',
+    fontWeight: {
+      regular: 400,
+      medium: 500,
+      strong: 600,
+    },
   },
   color: {
     foreground: {
